requirements.txt
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
openai
python-dotenv
schedule
requests
